Explicitly specify features needed for syn
Requires switch to resolver = "2", so that features for
dev-dependencies and regular dependencies are seperate.

Otherwise, our dev-dependency on trybuild would transitively
include the default syn features.

It appears the "printing" dependency is not directly required,
it is only needed by the tests.
However, I have added it because it's used for the 'Spanned' trait,
which is likely needed in the future.

Also, we already have a dependency on the 'quote' crate,
so the additional "printing" feature doesn't pull in much more.

Fixes issue #4